Convert tests to use TestItemRunner
